kinesis-producer-py/requirements.txt
mimesis
boto3
avro-python3